# src/main/resources/APIDOC.YAML
openapi: 3.0.1
info:
  title: Constructa OpenApi specification
  description: OpenApi documentation for Constructa BE
  contact:
    name: Ondřej Bureš
    url: 'https://digitalnoodles.cz/'
    email: bureson2@fel.cvut.cz
  version: '1.0'
servers:
  - url: 'http://localhost:8080/'
    description: Local ENV
security:
  - bearerAuth: []
paths:
...
  /api/v1/vehicles:
    get:
      tags:
        - Vehicles
      operationId: getVehicles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Vehicles
      operationId: updateVehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Vehicles
      operationId: createVehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDTO'
      security:
        - bearerAuth: []
  /api/v1/vehicles/reports:
    get:
      tags:
        - Vehicle reports
      operationId: getVehicleReports
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Vehicle reports
      operationId: updateVehicleReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleReportRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Vehicle reports
      operationId: createVehicleReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleReportRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/vehicles/{vehicleId}':
    get:
      tags:
        - Vehicles
      operationId: getVehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Vehicles
      operationId: deleteVehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
  '/api/v1/vehicles/{carId}/reports':
    get:
      tags:
        - Vehicle reports
      operationId: getVehicleReportsByVehicle
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
  /api/v1/vehicles/search:
    get:
      tags:
        - Vehicles
      operationId: getInputVehicles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleInputDTO'
      security:
        - bearerAuth: []
  '/api/v1/vehicles/reports/{vehicleReportId}':
    get:
      tags:
        - Vehicle reports
      operationId: getVehicleReport
      parameters:
        - name: vehicleReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/vehicles/reports/{reportId}':
    delete:
      tags:
        - Vehicle reports
      operationId: deleteVehicleReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
components:
  schemas:
...
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        name:
          type: string
        country:
          type: string
        postCode:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        titleBeforeName:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        titleAfterName:
          type: string
        bankAccount:
          type: string
        dateOfAcceptance:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        birthId:
          type: string
        hourRate:
          type: integer
          format: int32
        monthSalary:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/LocationDTO'
...
    VehicleRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        factory:
          type: string
        name:
          type: string
        registrationNumber:
          type: string
        conditionMotorcycleWatch:
          type: number
          format: double
        mileage:
          type: number
          format: double
        type:
          type: string
          enum:
            - CAR
            - VEHICLE
            - TRAILER
        boughtAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        qrCode:
          type: string
        vinCode:
          type: string
    VehicleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        factory:
          type: string
        name:
          type: string
        registrationNumber:
          type: string
        conditionMotorcycleWatch:
          type: number
          format: double
        mileage:
          type: number
          format: double
        type:
          type: string
        boughtAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        qrCode:
          type: string
        vinCode:
          type: string
    VehicleReportRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        originalConditionMotorcycleWatch:
          type: number
          format: double
        afterworkConditionMotorcycleWatch:
          type: number
          format: double
        cargoMass:
          type: integer
          format: int32
        cargoType:
          type: string
        distance:
          type: number
          format: double
        purchaseOfFuelLitres:
          type: number
          format: double
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        description:
          type: string
        vehicle:
          type: integer
          format: int64
        driver:
          type: integer
          format: int64
    VehicleReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        originalConditionMotorcycleWatch:
          type: number
          format: double
        afterworkConditionMotorcycleWatch:
          type: number
          format: double
        cargoMass:
          type: integer
          format: int32
        cargoType:
          type: string
        distance:
          type: number
          format: double
        purchaseOfFuelLitres:
          type: number
          format: double
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        description:
          type: string
        vehicle:
          $ref: '#/components/schemas/VehicleDTO'
        driver:
          $ref: '#/components/schemas/UserDTO'
...
    VehicleInputDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        factory:
          type: string
        name:
          type: string
        registrationNumber:
          type: string
        type:
          type: string
          enum:
            - CAR
            - VEHICLE
            - TRAILER
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT auth description
      in: header
      scheme: bearer
      bearerFormat: JWT